shaders hot reload fixed

+++ Tofu3D/Source/EditorResourcesAssetsWatcher.cs
@@ -3,7 +3,6 @@ using System.Linq;
 
 namespace TofuEngine;
 
-
 // Reflects changes done to shaders in this project EditorResources folder to bin folder where the base shaders are stored
 public class EditorResourcesAssetsWatcher
 {
@@ -12,8 +11,8 @@ public class EditorResourcesAssetsWatcher
 
     public void StartWatching()
     {
-        var a= AppContext.BaseDirectory;
-        _editorResourcesShadersAppContextPath = Path.Combine(a, "../../../EditorResources/Shaders");
+        var a = AppContext.BaseDirectory;
+        _editorResourcesShadersAppContextPath = Path.GetFullPath(Path.Combine(a, "../../../EditorResources/Shaders"));
         _watcher = new FileSystemWatcher(_editorResourcesShadersAppContextPath);
         _watcher.IncludeSubdirectories = true;
         _watcher.NotifyFilter = NotifyFilters.LastWrite; // | NotifyFilters.Size | NotifyFilters.LastAccess |
@@ -35,23 +34,31 @@ public class EditorResourcesAssetsWatcher
             return;
         }
 
-        string engineResourcesShaderRelativePath = Path.GetRelativePath(_editorResourcesShadersAppContextPath, e.FullPath);
+        var eFullPath = Path.GetFullPath(e.FullPath);
+        string engineResourcesShaderRelativePath =
+            Path.GetRelativePath(_editorResourcesShadersAppContextPath, eFullPath);
         IEnumerable<string> shadersInBinFolder = Directory.EnumerateFiles(Folders.EngineResourcesShaders);
         foreach (string shaderInBinFolder in shadersInBinFolder)
         {
             string binShaderRelativePath = Path.GetRelativePath(Folders.EngineResourcesShaders, shaderInBinFolder);
             if (engineResourcesShaderRelativePath.Equals(binShaderRelativePath, StringComparison.Ordinal))
             {
-                var a = Path.DirectorySeparatorChar+Path.GetRelativePath("/", e.FullPath);
+                var a = Path.DirectorySeparatorChar + Path.GetRelativePath("/", eFullPath);
 
-             File.Delete(shaderInBinFolder);
-                    File.Copy(a, shaderInBinFolder, overwrite: false);
+                if (Path.Exists(a) == false)
+                {
+                    continue;
+                }
+
+                _watcher.EnableRaisingEvents = false;
+                File.Delete(shaderInBinFolder);
+                File.Copy(a, shaderInBinFolder, overwrite: false);
+                _watcher.EnableRaisingEvents = true;
 
                 Tofu.ShaderManager.QueueShaderReload(shaderInBinFolder);
 
                 break;
             }
         }
-
     }
 }
